requirements.txt
# Swin Transformer 超声TOF预测项目依赖包
# 安装命令: pip install -r requirements.txt

# 深度学习框架
torch>=1.12.0,<2.0.0
torchvision>=0.13.0,<0.15.0
tensorboard>=2.10.0

# 科学计算和数据处理
numpy>=1.21.0,<1.25.0
pandas>=1.4.0,<2.0.0
scipy>=1.8.0,<1.11.0

# 机器学习
scikit-learn>=1.1.0,<1.3.0

# 图像处理
Pillow>=9.0.0,<10.0.0

# 可视化
matplotlib>=3.5.0,<3.7.0

# 进度条
tqdm>=4.64.0

# 模型可视化
torchviz>=0.0.2

# 深度学习模型库
timm>=0.6.0,<0.9.0

# 可选依赖（用于特定功能）
# 如果需要GPU加速，请确保安装对应CUDA版本的PyTorch
# 例如: torch==1.13.1+cu117 -f https://download.pytorch.org/whl/torch_stable.html
